db_mysql/requirements.txt
mysql-connector-python
pandas
tqdm
loguru
python-dotenv